feat: add Open Beauty Facts Parquet export
